% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmap.R
\name{findmap}
\alias{findmap}
\title{Stamen map API.}
\usage{
findmap(name, zoom_l, map_type)
}
\arguments{
\item{name}{Location name.}

\item{zoom_l}{Zoom level of the map.}

\item{map_type}{Stamen map type.}
}
\value{
A list, stamen map with predefined zoom level and map type.
}
\description{
Enter a city name and display its stamen map.
}
\details{
Input takes name of a location and zoom level of the map, output will return the related Stamen map.
}
